+++ Changelog.md
@@ -21,6 +21,9 @@
     * [Pull request 2652](https://bitbucket.org/osrf/gazebo/pull-request/2652)
     * [Issue #863](https://bitbucket.org/osrf/gazebo/issues/863)
 
+1. Fix loading sdf with orthographic projection camera
+    * [Pull request 3098](https://bitbucket.org/osrf/gazebo/pull-request/3098)
+
 ## Gazebo 9.8.0 (2019-XX-XX)
 
 1. Fix kinematic loops for DART 6.7 and later